+++ src/igor/igor/resources/model/iModel_Old.cpp
@@ -2,7 +2,7 @@
 // (c) Copyright 2012-2023 by Martin Loga
 // see copyright notice in corresponding header file
 
-#include <igor/resources/model/iModel.h>
+#include <igor/resources/model/iModel_Old.h>
 #include <igor/scene/nodes/iNodeManager.h>
 
 #include <iaux/data/iaString.h>
@@ -12,14 +12,14 @@ using namespace iaux;
 #include <fstream>
 
 namespace igor
-{   
+{
 
-    iModel::iModel(const iParameters &parameters)
-        : iResource("model", parameters)
+    iModel_Old::iModel_Old(const iaString &name, iResourceCacheMode cacheMode, iModelDataInputParameterPtr parameter)
+        : _name(name), _parameter(parameter), _cacheMode(cacheMode)
     {
     }
 
-    iModel::~iModel()
+    iModel_Old::~iModel_Old()
     {
         if (nullptr != _node)
         {
@@ -27,7 +27,32 @@ namespace igor
         }
     }
 
-    void iModel::setNode(iNodePtr node)
+    iResourceCacheMode iModel_Old::getCacheMode() const
+    {
+        return _cacheMode;
+    }
+
+    iModelDataInputParameterPtr iModel_Old::getParameters() const
+    {
+        return _parameter;
+    }
+
+    const iaString &iModel_Old::getName() const
+    {
+        return _name;
+    }
+
+    iModelState iModel_Old::getState()
+    {
+        return _state;
+    }
+
+    void iModel_Old::setState(iModelState state)
+    {
+        _state = state;
+    }
+
+    void iModel_Old::setNode(iNodePtr node)
     {
         con_assert(node != nullptr, "zero pointer");
         con_assert(_node == nullptr, "can't override data");
@@ -39,7 +64,7 @@ namespace igor
         }
     }
 
-    iNodePtr iModel::getNodeCopy()
+    iNodePtr iModel_Old::getNodeCopy()
     {
         con_assert(_node != nullptr, "zero pointer. data not ready");
 
@@ -53,7 +78,7 @@ namespace igor
         }
     }
 
-    iNodePtr iModel::getNode()
+    iNodePtr iModel_Old::getNode()
     {
         return _node;
     }