+++ src/igor/igor/resources/model/loader/iModelDataIOOBJ.h
@@ -133,7 +133,7 @@ namespace igor
         \param filename filename of file to load
         \return pointer to node tree
         */
-        virtual iNodePtr importData(const iaString &filename, iModelDataInputParameterPtr parameter);
+        virtual iNodePtr importData(const iaString &filename, iModelDataInputParameterPtr parameter) override;
 
         /*! creates an instance of this class